requirements.txt
attr==0.3.2
ConfigParser==7.0.0
contextlib2==21.6.0
cryptography==42.0.7
Cython==3.0.10
dl==0.1.0
docutils==0.21.2
fastapi==0.111.0
filelock==3.14.0
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.2
Jinja2==3.1.2
jnius==1.1.0
keyring==25.2.0
langchain==0.1.17
mock==5.1.0
numpy==1.26.4
Pillow==10.3.0
pinecone==4.0.0
protobuf==5.26.1
pyOpenSSL==24.1.0
python-dotenv==1.0.1
pywin32==306
railroad==0.5.0
redis==5.0.4
Sphinx==7.3.7
thread==2.0.1
tornado==6.4
trove_classifiers==2024.4.10
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
